+++ neoforge/src/main/java/dev/boxadactle/coordinatesdisplay/neoforge/CoordinatesDisplayNeoforge.java
@@ -93,6 +93,7 @@ public class CoordinatesDisplayNeoforge {
             e.register(Bindings.copyPosTp);
             e.register(Bindings.changeHudPosition);
             e.register(Bindings.cycleDisplayMode);
+            e.register(Bindings.toggle3DCompass);
         }
     }