+++ CHANGELOG.md
@@ -1,2 +1,7 @@
 ## Changes
-- Added forge support
+- Added 3d compass rendering to config gui
+- Added toggle3dCompass subcommand
+- Added 3d compass toggle keybind
+
+## New Features
+- It is now possible to render a 3d compass client-side in the world